style: rename context | [#267]
