+++ Electron/src/__tests__/pages/Explorar.test.tsx
@@ -5,7 +5,7 @@ import Explorar from 'pages/Explorar/Explorar';
 import { BrowserRouter } from 'react-router-dom';
 import Global from 'global/global';
 import getMockHeaders from 'utils/mockHeaders';
-import { SongNameChangeContextProvider } from 'hooks/useSongChangeContextApi';
+import { NowPlayingContextProvider } from 'providers/NowPlayingProvider';
 
 const playlistName = 'playlisttest';
 const songName = 'songName';
@@ -84,9 +84,9 @@ test('Render Explorar and get Genres', async () => {
   const component = await act(() => {
     return render(
       <BrowserRouter>
-        <SongNameChangeContextProvider>
+        <NowPlayingContextProvider>
           <Explorar refreshSidebarData={jest.fn()} />
-        </SongNameChangeContextProvider>
+        </NowPlayingContextProvider>
       </BrowserRouter>,
     );
   });
@@ -97,9 +97,9 @@ test('Explorar filter by name', async () => {
   const component = await act(() => {
     return render(
       <BrowserRouter>
-        <SongNameChangeContextProvider>
+        <NowPlayingContextProvider>
           <Explorar refreshSidebarData={jest.fn()} />
-        </SongNameChangeContextProvider>
+        </NowPlayingContextProvider>
       </BrowserRouter>,
     );
   });
