+++ Electron/src/__tests__/pages/PlaylistArtist.test.tsx
@@ -8,7 +8,7 @@ import UserType from 'utils/role';
 import getMockHeaders from 'utils/mockHeaders';
 
 import * as TokenModule from 'utils/token';
-import { SongNameChangeContextProvider } from 'hooks/useSongChangeContextApi';
+import { NowPlayingContextProvider } from 'providers/NowPlayingProvider';
 
 const userName = 'prueba';
 const roleArtist = UserType.ARTIST;
@@ -90,14 +90,14 @@ test('Playlist artist role get all info', async () => {
   const component = await act(() => {
     return render(
       <MemoryRouter initialEntries={[`/playlist/${playlistDTOMockFetch.name}`]}>
-        <SongNameChangeContextProvider>
+        <NowPlayingContextProvider>
           <Routes>
             <Route
               path="/playlist/:id"
               element={<Playlist refreshSidebarData={jest.fn()} />}
             />
           </Routes>
-        </SongNameChangeContextProvider>
+        </NowPlayingContextProvider>
       </MemoryRouter>,
     );
   });
