+++ Electron/src/__tests__/pages/PlaylistUser.test.tsx
@@ -7,7 +7,7 @@ import Global from 'global/global';
 import UserType from 'utils/role';
 import getMockHeaders from 'utils/mockHeaders';
 import * as TokenModule from 'utils/token';
-import { SongNameChangeContextProvider } from 'hooks/useSongChangeContextApi';
+import { NowPlayingContextProvider } from 'providers/NowPlayingProvider';
 
 const userName = 'prueba';
 const roleUser = UserType.USER;
@@ -110,14 +110,14 @@ test('Playlist user role get all info', async () => {
   const component = await act(() => {
     return render(
       <MemoryRouter initialEntries={[`/playlist/${playlistDTOMockFetch.name}`]}>
-        <SongNameChangeContextProvider>
+        <NowPlayingContextProvider>
           <Routes>
             <Route
               path="/playlist/:id"
               element={<Playlist refreshSidebarData={jest.fn()} />}
             />
           </Routes>
-        </SongNameChangeContextProvider>
+        </NowPlayingContextProvider>
       </MemoryRouter>,
     );
   });
@@ -220,14 +220,14 @@ test('Playlist user role hit like button', async () => {
   const component = await act(() => {
     return render(
       <MemoryRouter initialEntries={[`/playlist/${playlistDTOMockFetch.name}`]}>
-        <SongNameChangeContextProvider>
+        <NowPlayingContextProvider>
           <Routes>
             <Route
               path="/playlist/:id"
               element={<Playlist refreshSidebarData={jest.fn()} />}
             />
           </Routes>
-        </SongNameChangeContextProvider>
+        </NowPlayingContextProvider>
       </MemoryRouter>,
     );
   });
@@ -339,14 +339,14 @@ test('Playlist user role get unlike button', async () => {
   const component = await act(() => {
     return render(
       <MemoryRouter initialEntries={[`/playlist/${playlistDTOMockFetch.name}`]}>
-        <SongNameChangeContextProvider>
+        <NowPlayingContextProvider>
           <Routes>
             <Route
               path="/playlist/:id"
               element={<Playlist refreshSidebarData={jest.fn()} />}
             />
           </Routes>
-        </SongNameChangeContextProvider>
+        </NowPlayingContextProvider>
       </MemoryRouter>,
     );
   });
@@ -472,14 +472,14 @@ test('Playlist user role update playlist', async () => {
   const component = await act(() => {
     return render(
       <MemoryRouter initialEntries={[`/playlist/${playlistDTOMockFetch.name}`]}>
-        <SongNameChangeContextProvider>
+        <NowPlayingContextProvider>
           <Routes>
             <Route
               path="/playlist/:id"
               element={<Playlist refreshSidebarData={refreshSidebarData} />}
             />
           </Routes>
-        </SongNameChangeContextProvider>
+        </NowPlayingContextProvider>
       </MemoryRouter>,
     );
   });
