+++ Electron/src/components/Cards/SongCard/SongCard.tsx
@@ -3,7 +3,7 @@ import ContextMenuSong from 'components/AdvancedUIComponents/ContextMenu/Song/Co
 import Popover, { PopoverPosition } from '@mui/material/Popover';
 import { useNavigate } from 'react-router-dom';
 import { PropsSongCard } from 'types/song';
-import { useSongNameChangeContext } from 'hooks/useSongChangeContextApi';
+import { useNowPlayingContext } from 'hooks/useNowPlayingContext';
 import styles from '../cards.module.css';
 import defaultThumbnailPlaylist from '../../../assets/imgs/DefaultThumbnailPlaylist.jpg';
 
@@ -13,7 +13,7 @@ export default function SongCard({
   photo,
   refreshSidebarData,
 }: PropsSongCard) {
-  const { changeSongName } = useSongNameChangeContext();
+  const { changeSongName } = useNowPlayingContext();
 
   const navigate = useNavigate();
